Changed length field from int to long.
Rev: server/java/src/se/idonex/roxen/RoxenResponse.java:1.2
Rev: server/java/src/se/idonex/roxen/RoxenStringResponse.java:1.3

+++ server/java/src/se/idonex/roxen/RoxenResponse.java
@@ -1,5 +1,5 @@
 /*
- * $Id: RoxenResponse.java,v 1.1 1999/12/19 00:26:01 marcus Exp $
+ * $Id: RoxenResponse.java,v 1.2 1999/12/21 00:06:39 marcus Exp $
  *
  */
 
@@ -9,9 +9,9 @@ public abstract class RoxenResponse {
 
   int errno;
   String type;
-  int len;
+  long len;
 
-  RoxenResponse(int _errno, String _type, int _len)
+  RoxenResponse(int _errno, String _type, long _len)
   {
     errno = _errno;
     type = _type;
