+++ server/java/src/se/idonex/roxen/RoxenStringResponse.java
@@ -1,5 +1,5 @@
 /*
- * $Id: RoxenStringResponse.java,v 1.2 1999/12/19 21:00:51 marcus Exp $
+ * $Id: RoxenStringResponse.java,v 1.3 1999/12/21 00:06:39 marcus Exp $
  *
  */
 
@@ -9,7 +9,7 @@ public class RoxenStringResponse extends RoxenResponse {
 
   String data;
 
-  RoxenStringResponse(int _errno, String _type, int _len, String _data)
+  RoxenStringResponse(int _errno, String _type, long _len, String _data)
   {
     super(_errno, _type, _len);
     data = _data;
